Moved gui code
