+++ src/simulator/views/gui/common.py
